requirements.txt
asgiref==3.2.10
Django==3.1.2
django-allauth==0.42.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.3.1